\docType{data}
\name{Oil}
\alias{Oil}
\title{Oil paml yield}
\source{
Aldo M. Garay, Victor H. Lachos, Carlos A. Abanto-Valle
(2011). "Nonlinear regression models based on scale mixture
of skew-normal distributions". _Journal of the Korean
Stastical Society_, 40, 115-124.
}
\usage{
data(Oil)
}
\description{
Growth and yield of palm oil
}

