\docType{data}
\name{Ultrasonic}
\alias{Ultrasonic}
\title{Ultrasonic Calibration}
\format{A data frame with 214 observations with y as the ultrasonic measuraments and x the metal distance}
\source{
Victor H. Lachos, Dipankar Bandyopadhyay and Aldo M. Garay
(2011). "Heteroscedastic nonlinear regression models based
on scale mixture of skew-normal distributions". _Statistics
and Probability Letters_, 81, 1208-1217.
}
\usage{
data(Ultrasonic)
}
\description{
The data is a result of a ultrasonic calibration study
perfomed by National Institute of Standard and Technology.
}

